+++ unit-tests/forms/disable-save/security-groups.test.js
@@ -1,5 +1,5 @@
 const { assert } = require("chai");
-const { disableSave } = require("../../../client/src/lib");
+const { disableSave, state } = require("../../../client/src/lib");
 
 describe("security groups", () => {
   it("should return true if a security group has an invalid name", () => {
@@ -10,45 +10,7 @@ describe("security groups", () => {
           name: "@@@",
         },
         {
-          craig: {
-            store: {
-              json: {
-                security_groups: [
-                  {
-                    name: "frog",
-                  },
-                  {
-                    name: "toad",
-                  },
-                ],
-                vpcs: [
-                  {
-                    name: "frog",
-                    acls: [
-                      {
-                        name: "frog",
-                        rules: [
-                          {
-                            name: "frog",
-                          },
-                          {
-                            name: "mmm",
-                          },
-                        ],
-                      },
-                      {
-                        name: "aaa",
-                      },
-                    ],
-                  },
-                  {
-                    name: "toad",
-                    acls: [],
-                  },
-                ],
-              },
-            },
-          },
+          craig: state(),
           data: {
             name: "frog",
           },
@@ -66,45 +28,7 @@ describe("security groups", () => {
           resource_group: null,
         },
         {
-          craig: {
-            store: {
-              json: {
-                security_groups: [
-                  {
-                    name: "frog",
-                  },
-                  {
-                    name: "toad",
-                  },
-                ],
-                vpcs: [
-                  {
-                    name: "frog",
-                    acls: [
-                      {
-                        name: "frog",
-                        rules: [
-                          {
-                            name: "frog",
-                          },
-                          {
-                            name: "mmm",
-                          },
-                        ],
-                      },
-                      {
-                        name: "aaa",
-                      },
-                    ],
-                  },
-                  {
-                    name: "toad",
-                    acls: [],
-                  },
-                ],
-              },
-            },
-          },
+          craig: state(),
           data: {
             name: "frog",
           },
@@ -123,45 +47,7 @@ describe("security groups", () => {
           vpc: null,
         },
         {
-          craig: {
-            store: {
-              json: {
-                security_groups: [
-                  {
-                    name: "frog",
-                  },
-                  {
-                    name: "toad",
-                  },
-                ],
-                vpcs: [
-                  {
-                    name: "frog",
-                    acls: [
-                      {
-                        name: "frog",
-                        rules: [
-                          {
-                            name: "frog",
-                          },
-                          {
-                            name: "mmm",
-                          },
-                        ],
-                      },
-                      {
-                        name: "aaa",
-                      },
-                    ],
-                  },
-                  {
-                    name: "toad",
-                    acls: [],
-                  },
-                ],
-              },
-            },
-          },
+          craig: state(),
           data: {
             name: "frog",
           },
@@ -170,50 +56,56 @@ describe("security groups", () => {
       "it should be true"
     );
   });
-  it("should return true if security group rule has invalid name", () => {
-    assert.isTrue(
-      disableSave(
-        "sg_rules",
-        {
-          name: "@@@",
-        },
-        { innerFormProps: { rules: [] }, data: { name: "" } }
-      )
-    );
-  });
-  it("should return true if security group rule has invalid source", () => {
-    assert.isTrue(
-      disableSave(
-        "sg_rules",
-        { name: "aa", source: "mm" },
-        { innerFormProps: { rules: [] }, data: { name: "" } }
-      )
-    );
-  });
-  it("should return true if security group rule has invalid port", () => {
-    assert.isTrue(
-      disableSave(
-        "sg_rules",
-        {
-          name: "aa",
-          source: "1.2.3.4",
-          ruleProtocol: "udp",
-          rule: {
-            port_min: -1,
-            port_max: null,
+  describe("rules", () => {
+    it("should return true if security group rule has invalid name", () => {
+      assert.isTrue(
+        disableSave(
+          "sg_rules",
+          {
+            name: "@@@",
           },
-        },
-        { innerFormProps: { rules: [{ name: "ff" }] }, data: { name: "aa" } }
-      )
-    );
-  });
-  it("should return true if security group rule has invalid name in modal", () => {
-    assert.isTrue(
-      disableSave(
-        "sg_rules",
-        { name: "@@@" },
-        { rules: [], data: { name: "" }, isModal: true }
-      )
-    );
+          { innerFormProps: { rules: [] }, data: { name: "" }, craig: state() }
+        )
+      );
+    });
+    it("should return true if security group rule has invalid source", () => {
+      assert.isTrue(
+        disableSave(
+          "sg_rules",
+          { name: "aa", source: "mm" },
+          { innerFormProps: { rules: [] }, data: { name: "" }, craig: state() }
+        )
+      );
+    });
+    it("should return true if security group rule has invalid port", () => {
+      assert.isTrue(
+        disableSave(
+          "sg_rules",
+          {
+            name: "aa",
+            source: "1.2.3.4",
+            ruleProtocol: "udp",
+            rule: {
+              port_min: -1,
+              port_max: null,
+            },
+          },
+          {
+            innerFormProps: { rules: [{ name: "ff" }] },
+            data: { name: "aa" },
+            craig: state(),
+          }
+        )
+      );
+    });
+    it("should return true if security group rule has invalid name in modal", () => {
+      assert.isTrue(
+        disableSave(
+          "sg_rules",
+          { name: "@@@" },
+          { rules: [], data: { name: "" }, isModal: true, craig: state() }
+        )
+      );
+    });
   });
 });