+++ CHANGELOG.md
@@ -6,12 +6,14 @@ All notable changes to this project will be documented in this file.
 
 ### Upgrade Notes
 
+- Transit Gateway VPC CRNs have been removed from the Transit Gateway form. Existing `craig.json` files with this field will still render normally. Please reach out to us if the GUI functionality is needed for your environment
 
 ### Features
 
 - Users can now edit the size of SAP log volumes for Power VS
 - When using the search bar in the hamburger menu, results will be returned based on documentation text
 - Users can now choose a router hostname from an existing Classic VLAN on the Classic VLAN page
+- Users can now import an existing Transit Gateway
 
 ### Fixes
 
@@ -26,6 +28,7 @@ All notable changes to this project will be documented in this file.
 - Fixed an issue causing Classic Infrastructure resources to not be added to CRAIG output archive
 - Fixed an issue allowing users to create invalid Classic VLAN names with more than 20 characters
 - Fixed an issue causing Subnets page to crash when switching projects
+- Fixed an issue allowing VPCs and Power VS Workspaces to be connected to more than one local or global Transit Gateway
 
 ## 1.6.0